+++ src/main/java/amidst/mojangapi/world/filter/WorldFilter_MatchAll.java
@@ -9,7 +9,7 @@ import amidst.mojangapi.world.World;
 public class WorldFilter_MatchAll extends WorldFilter {
 	private final List<WorldFilter> filters;
 
-	public WorldFilter_MatchAll(long worldFilterSize, List<WorldFilter> filters) {
+	public WorldFilter_MatchAll(int worldFilterSize, List<WorldFilter> filters) {
 		super(worldFilterSize);
 		this.filters = filters;
 	}
